refine VectorMemoryCollection refactor
